Move Table stuff into its own folder
