+++ app/javascript/Tables/GraphQLReactTableWrapper.jsx
