+++ app/javascript/Tables/TableUtils.js
